% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCohortProgression.R
\name{calcCohortProgression}
\alias{calcCohortProgression}
\title{Calculate the cohort progression}
\usage{
calcCohortProgression(configInfo, nmInfo, mltInfo, grwInfo, m2mInfo,
  recInfo, input, verbose = FALSE)
}
\arguments{
\item{configInfo}{- list created by \code{setConfigInfo} with model configuration elements}

\item{nmInfo}{- list created by \code{natmortServer} with model natural mortality rates elements}

\item{mltInfo}{- list created by \code{moltServer} with model molt probability elements}

\item{grwInfo}{- list created by \code{growthServer} with growth transition elements}

\item{m2mInfo}{- list created by \code{m2mServer} with prM2M elements}

\item{recInfo}{- list created by \code{recServer} with recruitment distribution elements}

\item{input}{- list with cohort progression parameters and other information}

\item{verbose}{- flag (T/F) to print debugging info}
}
\value{
list - list with n_msz as a vector and as a dataframe (see \link{ @details })
}
\description{
Function to calculate the cohort progressionn.
}
\details{
Returns a list with elements:<cr>
<ul>
 <li> params - a list with input parameter values </li>
 <li> n_ymsz - an array with the cohort progression </li>
 <li> mdfrN_ymsz - a dataframe wtih the cohort progression </li>
 <li> p_ymsz - an array with the normalized cohort progression </li>
 <li> mdfrP_ymsz - a dataframe wtih the normalized cohort progression </li>
 <li> n_msz - an array with the equilibrium size distribution </li>
 <li> mdfrN_msz - a dataframe wtih the equilibrium size distribution </li>
 <li> p_msz - an array with the normalized equilibrium size distribution </li>
 <li> mdfrP_msz - a dataframe wtih the normalized equilibrium size distribution </li>
</ul>

The \code{input} list must have the following elements:<cr>
<ul>
 <li> maxAge - max age (years) to include in progression </li>
</ul>
}
