% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowth.R
\name{calcGrowth}
\alias{calcGrowth}
\title{Calculate mean growth and the growth transition matrix}
\usage{
calcGrowth(configInfo, input, verbose = FALSE)
}
\arguments{
\item{configInfo}{- list created by \code{setConfigInfo} with model configuration elements}

\item{input}{- list with growth parameters and other information}
}
\value{
list - list with mean growth and growth transition matrix (see \link{ @details })
}
\description{
Function to calculate mean growth and the growth transition matrix.
}
\details{
Returns a list with elements:<cr>
<ul>
 <li> params - a list with input parameter values </li>
 <li> mdfrMnGrowth - a dataframe with mean growth by pre-molt size </li>
 <li> mdfrPrGrowth - a dataframe wtih probability of growth into a post-molt size by pre-molt size </li>
 <li> prGr_zz - the size-transition matrix (rows: pre-molt size bins; cols: post-molt size bins)
</ul>

The \code{input} list has elements:<cr>
<ul>
 <li> zA - reference pre-molt size (mm CW) for pA</li>
 <li> pA - parameter reflecting mean post-molt size at pre-molt size zA</li>
 <li> zB - reference pre-molt size (mm CW) for pB</li>
 <li> pB - parameter reflecting mean post-molt size at pre-molt size zB/li>
 <li> pBeta - parameter reflecting scale function for gamma distribution zB/li>
 <li> maxZBEx - parameter reflecting max extent of potential growth (in size bins) zB/li>
</ul>
}
