% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNaturalMortality.R
\name{calcNaturalMortality}
\alias{calcNaturalMortality}
\title{Calculate natural mortality rates}
\usage{
calcNaturalMortality(configInfo, input, verbose = FALSE)
}
\arguments{
\item{configInfo}{- list created by \code{setConfigInfo} with model configuration elements}

\item{input}{- list with natural mortality rate parameters and other information}
}
\value{
list - list with M_ms as a matrix and as a dataframe (see \link{ @details })
}
\description{
Function to calculate natural mortality rates.
}
\details{
Returns a list with elements:<cr>
<ul>
 <li> params - a list with input parameter values </li>
 <li> M_ms - a matrix of natural mortality rates indexed by maturity state and shell condition </li>
 <li> mdfrM_ms - a dataframe wtih the the same information </li>
</ul>

The \code{input} list has elements:<cr>
<ul>
 <li> pM - base natural mortality </li>
 <li> pDM1 - multiplier for immature crab </li>
 <li> pDM2 - multiplier for mature crab </li>
</ul>
}
