% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrM2M.R
\name{calcPrM2M}
\alias{calcPrM2M}
\title{Calculate mean growth and the growth transition matrix}
\usage{
calcPrM2M(configInfo, input, verbose = FALSE)
}
\arguments{
\item{configInfo}{- list created by \code{setConfigInfo} with model configuration elements}

\item{input}{- list with growth parameters and other information}
}
\value{
list - list with prM2M as a vector and as a dataframe (see \link{ @details })
}
\description{
Function to calculate mean growth and the growth transition matrix.
}
\details{
Returns a list with elements:<cr>
<ul>
 <li> params - a list with input parameter values </li>
 <li> prM2M_z - a vector with  wtih probability of molt-to-maturity by pre-molt size </li>
 <li> mdfrPrM2M - a dataaframe wtih probability of molt-to-maturity by pre-molt size </li>
</ul>

The \code{input} list has elements:<cr>
<ul>
 <li> z50 - location parameter for parametric curve (logistic; size at inflection point) </li>
 <li> b50 - scale parameter for parametric curve (logistic; width) </li>
 <li> z0  - min size at which prM2M becomes non-zero (using parametric description) </li>
 <li> pLgtScl - logit-scale parameters reflecting non-parametric curve for prM2M </li>
 <li> n0  - number of initial size bins at which prM2M is zero (using non-parametric curve description) </li>
 <li> useNonParametric - flag (T/F) to use non-parametric curve for prM2M </li>
</ul>
}
