% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrMolt.R
\name{calcPrMolt}
\alias{calcPrMolt}
\title{Calculate probability at size of molting for immature crab}
\usage{
calcPrMolt(configInfo, input, verbose = FALSE)
}
\arguments{
\item{configInfo}{- list created by \code{setConfigInfo} with model configuration elements}

\item{input}{- list with growth parameters and other information}
}
\value{
list - list with prMolt as a vector and as a dataframe (see \link{ @details })
}
\description{
Function to calculate probability at size of molting for immature crab.
}
\details{
Returns a list with elements:<cr>
<ul>
 <li> params - a list with input parameter values </li>
 <li> prMolt_z - a vector with  wtih probability of annual molt by pre-molt size </li>
 <li> mdfrPrMolt - a dataframe wtih probability of annual molt by pre-molt size </li>
</ul>

The \code{input} list has elements:<cr>
<ul>
 <li> z50 - location parameter for logistic curve (size at inflection point) </li>
 <li> b50 - scale parameter for parametric curve </li>
 <li> min - mininimum motling probability </li>
 <li> z1  - minimum size at which prMolt is less than 1 </li>
</ul>
}
