% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrRecAtZ.R
\name{calcPrRecAtZ}
\alias{calcPrRecAtZ}
\title{Calculate the recruitment size distribution}
\usage{
calcPrRecAtZ(configInfo, input, verbose = FALSE)
}
\arguments{
\item{configInfo}{- list created by \code{setConfigInfo} with model configuration elements}

\item{input}{- list with recruitment size distribution parameters and other information}
}
\value{
list - list with prR as a vector and as a dataframe (see \link{ @details })
}
\description{
Function to calculate the recruitment size distribution.
}
\details{
Returns a list with elements:<cr>
<ul>
 <li> params - a list with input parameter values </li>
 <li> prR_z - a vector with the recruitment size distribution </li>
 <li> mdfrPrRecAtZ - a dataframe wtih the recruitment size distribution </li>
</ul>

The \code{input} list has elements:<cr>
<ul>
 <li> alpha - location parameter for gamma distribution </li>
 <li> beta - scale parameter for gamma distribution </li>
 <li> mxZ - max size at recruitment </li>
</ul>
}
