% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortServer.R
\name{cohortServer}
\alias{cohortServer}
\title{Server function for the Tanner crab model cohort progression Shiny module}
\usage{
cohortServer(input, output, configInfo, nmResults, mltResults, grwResults,
  m2mResults, recResults, session)
}
\arguments{
\item{input}{- the usual shiny input variable}

\item{output}{- the usual shiny output variable}

\item{configInfo}{- reactive variable yielding a list with model configuration info}

\item{session}{- the usual shiny session variable}
}
\value{
reactive object which returns a list with elements
<ul>
  <li> cpInfo -  a list with elements inpInfo, prR_z, and mdfrPrRecAtZ (see \code{\link{calcCohortProgression}})</li>
  <li> pltPrRecAtZ - a ggplot2 object
</ul>
}
\description{
Server function for the Tanner crab model cohort progressionn Shiny module.
}
\details{
Updates the parameters, results, and associated ggplot2 objects
associated with the Tanner crab cohort progression model.
}
