% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortUI.R
\name{cohortUI}
\alias{cohortUI}
\title{UI for the Shiny Tanner crab recruitment cohort progression module}
\usage{
cohortUI(id)
}
\arguments{
\item{id}{- a character string that uniquely identifies the cohort progression module in the enclosing Shiny UI.}
}
\value{
A Shiny tabPanel allowing the user to change the cohort progression parameters and plot the results.
}
\description{
Function to create the UI for the Shiny Tanner crab cohort progression module.
}
\details{
Allows the user to change the cohort progression parameters and plot the results for the Tanner crab model.
}
