% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configUI.R
\name{configUI}
\alias{configUI}
\title{UI for model configuration ("config") Shiny module}
\usage{
configUI(id)
}
\arguments{
\item{id}{- a character string that uniquely identifies the model configuration module in the enclosing Shiny UI.}
}
\value{
A Shiny tabPanel allowing the user to change the model configuration.
}
\description{
Function to create the UI for the Tanner crab model configuration ("config") Shiny module
}
\details{
Allows the user to change bin size, min size, and max size for the Tanner crab model.
}
