% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqzServer.R
\name{eqzServer}
\alias{eqzServer}
\title{Server function for the Tanner crab model equilibrium size distribution Shiny module}
\usage{
eqzServer(input, output, configInfo, nmResults, mltResults, grwResults,
  m2mResults, recResults, cpResults, session)
}
\arguments{
\item{input}{- the usual shiny input variable}

\item{output}{- the usual shiny output variable}

\item{configInfo}{- reactive variable yielding a list with model configuration info}

\item{nmResults}{- natural mortality results list}

\item{mltResults}{- probability of molt results list}

\item{grwResults}{- growth results list}

\item{m2mResults}{- molt-to-maturity results list}

\item{recResults}{- recruitment results list}

\item{session}{- the usual shiny session variable}

\item{cpReults}{- cohort progression results list}
}
\value{
reactive object which returns a list with elements
<ul>
  <li> cpInfo -  a list with elements inpInfo, prR_z, and mdfrPrRecAtZ (see \code{\link{calcCohortProgression}})</li>
  <li> pltPrRecAtZ - a ggplot2 object
</ul>
}
\description{
Server function for the Tanner crab model equilibrium size distribution Shiny module.
}
\details{
Updates the parameters, results, and associated ggplot2 objects
associated with the Tanner crab cohort progression model.
}
