% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqzUI.R
\name{eqzUI}
\alias{eqzUI}
\title{UI for the Shiny Tanner crab equilibrium size distribution module}
\usage{
eqzUI(id)
}
\arguments{
\item{id}{- a character string that uniquely identifies the equilibrium size distribution module in the enclosing Shiny UI.}
}
\value{
A Shiny tabPanel allowing the user to change the equilibrium size distribution parameters and plot the results.
}
\description{
Function to create the UI for the Shiny Tanner crab equilibrium size distribution module.
}
\details{
Allows the user to change the equilibrium size distribution parameters and plot the results for the Tanner crab model.
}
