% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthServer.R
\name{growthServer}
\alias{growthServer}
\title{Server function for the Tanner crab model growth Shiny module}
\usage{
growthServer(input, output, configInfo, session)
}
\arguments{
\item{configInfo}{- reactive variable yielding a list with model configuration info}
}
\value{
a reactive object which returns a list with elements:<cr>
<ul>
 <li> growthInfo - a list returned from \code{\link{calcGrowth}} </li>
 <li> plotMnG - a ggplot2 object with mean growth by pre-molt size </li>
 <li> plotPrG - a ggplot2 object wtih probability of growth into a post-molt size by pre-molt size </li>
</ul>
}
\description{
Server function for the Tanner crab model growth Shiny module.
}
\details{
Updates the parameters, mean growth, growth transition matrix, and associated ggplot2 objects
associated with the Tanner crab growth model.
}
