% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthUI.R
\name{growthUI}
\alias{growthUI}
\title{UI for the Shiny Tanner crab growth module}
\usage{
growthUI(id)
}
\arguments{
\item{id}{- a character string that uniquely identifies the growth module in the enclosing Shiny UI.}
}
\value{
A Shiny tabPanel allowing the user to change growth parameters and plot mean growth and size transition probabilities.
}
\description{
Function to create the UI for the Shiny Tanner crab growth module.
}
\details{
Allows the user to change growth parameters and plot mean growth and size transition probabilities for the Tanner crab model.
}
