% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Get the shinyTC_CohortProgression app to run it}
\usage{
launchApp()
}
\value{
shiny application object (which is run using shiny::runApp(obj,...))
}
\description{
Function to get the shinyTC_CohortProgression app to run it.
}
\details{
To run the app from the commandline or in an RStudio session, use
shiny::runApp(shinyTC.CohortProgression::launchApp())
}
