% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2mServer.R
\name{m2mServer}
\alias{m2mServer}
\title{Server function for the Tanner crab model molt-to-maturity Shiny module}
\usage{
m2mServer(input, output, configInfo, session)
}
\arguments{
\item{input}{- the usual shiny input variable}

\item{output}{- the usual shiny output variable}

\item{configInfo}{- reactive variable yielding a list with model configuration info}

\item{session}{- the usual shiny session variable}
}
\value{
reactive object which returns a list with elements
<ul>
  <li> m2mInfo -  a list with elements inpInfo, prM2M_z, and mdfrPrM2M (see \code{\link{calcPrM2M}})</li>
  <li> plotPrM2M - a ggplot2 object
</ul>
}
\description{
Server function for the Tanner crab model molt-to-maturity Shiny module.
}
\details{
Updates the parameters, molt-to-maturity probabilities, and associated ggplot2 objects
associated with the Tanner crab molt-to-maturity model.
}
