% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2mUI.R
\name{m2mUI}
\alias{m2mUI}
\title{UI for the Shiny Tanner crab molt-to-maturity module}
\usage{
m2mUI(id)
}
\arguments{
\item{id}{- a character string that uniquely identifies the molt-to-maturity module in the enclosing Shiny UI.}
}
\value{
A Shiny tabPanel allowing the user to change molt-to-maturity parameters and plot molt-to-maturity probabilities.
}
\description{
Function to create the UI for the Shiny Tanner crab molt-to-maturity module.
}
\details{
Allows the user to change molt-to-maturity parameters and plot molt-to-maturity probabilities for the Tanner crab model.
}
