% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moltServer.R
\name{moltServer}
\alias{moltServer}
\title{Server function for the Tanner crab model annual molt probability Shiny module}
\usage{
moltServer(input, output, configInfo, session)
}
\arguments{
\item{configInfo}{- reactive variable yielding a list with model configuration info}
}
\value{
reactive object which returns a list with elements
<ul>
  <li> moltInfo -  a list with elements inpInfo, prMolt_z, and mdfrPrMolt (see \code{\link{calcPrMolt}})</li>
  <li> plotPrMolt - a ggplot2 object
</ul>
}
\description{
Server function for the Tanner crab model molt probability Shiny module.
}
\details{
Updates the parameters, molt probabilities, and associated ggplot2 objects
associated with the Tanner crab annual molt model.
}
