% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moltUI.R
\name{moltUI}
\alias{moltUI}
\title{UI for the Shiny Tanner crab probability of molt module}
\usage{
moltUI(id)
}
\arguments{
\item{id}{- a character string that uniquely identifies the probability of molt module in the enclosing Shiny UI.}
}
\value{
A Shiny tabPanel allowing the user to change probability of molt parameters and plot probabilities of molt.
}
\description{
Function to create the UI for the Shiny Tanner crab probability of molt module.
}
\details{
Allows the user to change probability of molt parameters and plot probabilities of molt for the Tanner crab model.
}
