% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natmortServer.R
\name{natmortServer}
\alias{natmortServer}
\title{Server function for the Tanner crab model natural mortality rate Shiny module}
\usage{
natmortServer(input, output, configInfo, session)
}
\arguments{
\item{configInfo}{- reactive variable yielding a list with model configuration info}
}
\value{
reactive object which returns a list with elements
<ul>
  <li> nmInfo -  a list with elements inpInfo, M_ms, and mdfrM_ms (see \code{\link{calcNaturalMortality}})</li>
  <li> plotNM - a ggplot2 object
</ul>
}
\description{
Server function for the Tanner crab model natural mortality rate Shiny module.
}
\details{
Updates the parameters, natural mortality rates, and associated ggplot2 objects
associated with the Tanner crab annual natural mortality rate model.
}
