% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natmortUI.R
\name{natmortUI}
\alias{natmortUI}
\title{UI for the Shiny Tanner crab natural mortality module}
\usage{
natmortUI(id)
}
\arguments{
\item{id}{- a character string that uniquely identifies the natural mortality module in the enclosing Shiny UI.}
}
\value{
A Shiny tabPanel allowing the user to change natural mortality parameters and plot the results.
}
\description{
Function to create the UI for the Shiny Tanner crab natural mortality module.
}
\details{
Allows the user to change natural mortality parameters and plot the reults for the Tanner crab model. Default
parameter values are from males from the 2018 assessment model.
}
