% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.EquilibriumSizeDist.R
\name{plotPop.EquilibriumSizeDist}
\alias{plotPop.EquilibriumSizeDist}
\title{Plot the equilibrium size distribution}
\usage{
plotPop.EquilibriumSizeDist(mdfr, plotType, scales = "free_y",
  showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{mdfr}{- melted data frame with equilibrium size distribution}
}
\value{
list of ggplot2 objects
}
\description{
Function to plot the equilibrium size distribution
}
\details{

}
