% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recServer.R
\name{recServer}
\alias{recServer}
\title{Server function for the Tanner crab model reccruitment size distribution Shiny module}
\usage{
recServer(input, output, configInfo, session)
}
\arguments{
\item{configInfo}{- reactive variable yielding a list with model configuration info}
}
\value{
reactive object which returns a list with elements
<ul>
  <li> recInfo -  a list with elements inpInfo, prR_z, and mdfrPrRecAtZ (see \code{\link{calcPrRecAtZ}})</li>
  <li> plotPrRecAtZ - a ggplot2 object
</ul>
}
\description{
Server function for the Tanner crab model reccruitment size distribution Shiny module.
}
\details{
Updates the parameters, recruitment size distribution, and associated ggplot2 objects
associated with the Tanner crab annual recruitment-at-size model.
}
