% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recUI.R
\name{recUI}
\alias{recUI}
\title{UI for the Shiny Tanner crab recruitment size distribution module}
\usage{
recUI(id)
}
\arguments{
\item{id}{- a character string that uniquely identifies the recruitment size distribution module in the enclosing Shiny UI.}
}
\value{
A Shiny tabPanel allowing the user to change recruitment size distribution parameters and plot the distribution.
}
\description{
Function to create the UI for the Shiny Tanner crab recruitment size distribution module.
}
\details{
Allows the user to change recruitment size distribution parameters and plot the distribution for the Tanner crab model.
}
