% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConfigInfo.R
\name{setConfigInfo}
\alias{setConfigInfo}
\title{Set the model configuration state}
\usage{
setConfigInfo(input, verbose = FALSE)
}
\value{
list - list with model configuration info (see \link{ @details })
}
\description{
Function to set the model configuration state.
}
\details{
Returns a list that reflects the
current model configuration state, with elements:<cr>
<ul>
 <li> moltTime - time of molting/mating (as fraction of year from July 1) </li>
 <li> maturityStates - maturity states </li>
 <li> shellConditions - shell conditions </li>
 <li> binZ - size bin width (mm CW) </li>
 <li> minZ - min size cutpoint (mm CW) </li>
 <li> maxZ - max size cutpoint (mm CW) </li>
 <li> zBs - size bins (mm CW) </li>
 <li> Zcs - size bin cutpoints (mm CW) </li>
</ul>

The input list has elements:<cr>
<ul>
 <li> moltTime - molt time (as fraction of year from July 1)  </li>
 <li> mss - text yielding a character vector of maturity states when parsed/evaluated  </li>
 <li> scs - text yielding a character vector of shell conditions when parsed/evaluated </li>
 <li> binZ - size bin width (mm CW) </li>
 <li> minZ - min size cutpoint (mm CW) </li>
 <li> maxZ - max size cutpoint (mm CW) </li>
</ul>
}
