% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI.R
\docType{data}
\name{shinyUI}
\alias{shinyUI}
\title{UI for shiny Cohort Progression app}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
shinyUI
}
\value{
the html UI for the app
}
\description{
UI for shiny Cohort Progression app.
}
\details{
Creates the UI for the app.
}
\keyword{datasets}
