% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch the shinyTC_CohortProgression app}
\usage{
launchApp()
}
\value{
shiny application object
}
\description{
Function to launch the shinyTC_CohortProgression app.
}
