% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyServer.R
\name{shinyServer}
\alias{shinyServer}
\title{Server for the shiny Cohort Progression app}
\usage{
shinyServer(input, output, session)
}
\arguments{
\item{input}{- the usual shiny input variable}

\item{output}{- the usual shiny output variable}

\item{session}{- the usual shiny session variable}
}
\value{
the html UI for the app
}
\description{
Server for the shiny Cohort Progression app.
}
\details{
Creates the UI for the app.
}
